two bugfixes: local pdcch_pdu in ul and wrong initialization place in find_candidate

+++ openair2/LAYER2/NR_MAC_gNB/mac_proto.h
@@ -260,13 +260,13 @@ NR_sched_pdcch_t set_pdcch_structure(gNB_MAC_INST *gNB_mac,
                                      NR_BWP_t *bwp,
                                      NR_Type0_PDCCH_CSS_config_t *type0_PDCCH_CSS_config);
 
-uint8_t find_pdcch_candidate(gNB_MAC_INST *mac,
-                             int cc_id,
-                             int aggregation,
-                             int nr_of_candidates,
-                             NR_sched_pdcch_t *pdcch,
-                             NR_ControlResourceSet_t *coreset,
-                             uint16_t Y);
+int find_pdcch_candidate(gNB_MAC_INST *mac,
+                         int cc_id,
+                         int aggregation,
+                         int nr_of_candidates,
+                         NR_sched_pdcch_t *pdcch,
+                         NR_ControlResourceSet_t *coreset,
+                         uint16_t Y);
 
 void fill_pdcch_vrb_map(gNB_MAC_INST *mac,
                         int CC_id,
